testing database
need to do/test updates

+++ 1-EV3/Unfinished/AppChessServer/Model/Model/Game.cs
@@ -4,9 +4,9 @@ namespace Model
 {
     public class GameDB
     {
-        public int codGame { get; set; }
-        public int codUserWhites { get; set; }
-        public int codUserBlacks { get; set; }
+        public long codGame { get; set; }
+        public long codUserWhites { get; set; }
+        public long codUserBlacks { get; set; }
         public string gameJson {  get; set; } = string.Empty;
     }
 }
